MP4Splitter: fix C4018 warnings. optimization for previous commit
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@3696 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/parser/MP4Splitter/Ap4AsyncReaderStream.cpp
@@ -31,7 +31,7 @@ AP4_Result AP4_AsyncReaderStream::Read(void* buffer, AP4_Size bytesToRead, AP4_S
 {
 	__int64 bytesAvail = m_pFile->GetRemaining();
 
-	if(bytesAvail < bytesToRead) {
+	if(bytesAvail < (long long)bytesToRead) {
 		if(bytesRead) {
 			*bytesRead = bytesAvail;
 		}
